requirements.txt
streamlit>=1.35.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
numpy>=1.24.0
streamlit-drawable-canvas>=0.9.3
